+++ gammapy/maps/core.py
@@ -9,7 +9,7 @@ from astropy.io import fits
 from gammapy.maps.hpx import HpxGeom
 from gammapy.utils.scripts import make_path
 from .geom import MapAxis, MapCoord, pix_tuple_to_idx
-from .utils import INVALID_VALUE, JsonQuantityDecoder
+from .utils import JsonQuantityDecoder
 
 __all__ = ["Map"]
